+++ src/Measure/DataBuffer.h
@@ -19,10 +19,10 @@ namespace MuzicAnalyser { namespace Measure
         DataBuffer(const uint16_t& numberOfMeasures, const uint8_t& channelsNumber, const uint16_t& offset);
         ~DataBuffer();
 
-        uint16_t*& GetChannelData(Channel channel);
-        Channel GetChannelNameByNumber(uint16_t channelNumber);
+        int16_t*& GetChannelData(Channel channel);
+        Channel GetChannelNameByNumber(uint8_t channelNumber);
 
-        uint16_t** data;
+        int16_t** data;
 
         uint16_t numberOfMeasures;
         uint8_t channelsNumber;
